requirements.txt
jinja2>=3.1
pyyaml>=6.0